Creating a tracklist
Scrapping the idea that each individual song gets a audio player.

New idea is to make a single audio player with a list of song selections and represent the content in the audio player and play the song after it has been selected.

Created the format for the track list. Must figure out the javascript for this.