Relocating Files
Moved the index html in the HTML folder and moved the song.html out of the songs folder, deleting the songs folder.

Deleting the songs folder is to create a dynamically rendered song.html file that renders different songs by the looking at the URL while having the same html file.