Adding EJS code
The front end ejs files will dynamically render the songs section of the index page suing a for loop with data sent in the songs variable from the backend.

Then the song.ejs file will be a template that can represent different songs by populating the page with different data.

The 404 error page must be an ejs page so the backend can use the render method on the response object.