Fixing Partials Path
Used relative path which appended to the current path. This made the server get pinged for paths that could not be found from the current path to relative path.

Correction is to use absolute path to access the express static files which are in the public folder. This was specified in the server.js code on line 16.