Server Update
Adding Routing for new songs. The song page is fed different variables from the ejs rendering.

The Index routing must only be fed the 4 elements from the json file, which are the latest songs.

The Discography routing is fed the entire songs array to be represented.

NOTE:
The songs.json file is temporary and the songs will be stored in a database in the future. The Discography page will also render more songs on the click of a button, requesting more songs from the Database. Although, at the moment there are not enough songs to have multiple calls to load new content.